saperate comments/reviews controller and model
