+++ models/reviews.model.js
@@ -16,17 +16,6 @@ const selectReviews = async () => {
   return rows;
 };
 
-const selectCommentsByReviewId = async (review_id) => {
-  const { rows: comments, rowCount: commentsCount } = await db.query(
-    `SELECT * FROM comments 
-     WHERE review_id = $1 
-     ORDER BY created_at DESC;`,
-    [review_id]
-  );
-
-  return comments;
-};
-
 const selectReviewById = async (review_id) => {
   const { rows, rowCount } = await db.query(
     `
@@ -40,4 +29,4 @@ const selectReviewById = async (review_id) => {
   return rows[0];
 };
 
-module.exports = { selectReviews, selectReviewById, selectCommentsByReviewId };
+module.exports = { selectReviews, selectReviewById };
